Added Player Controller Script
