+++ Through the Dungeon/.idea/.idea.Through the Dungeon/.idea/indexLayout.xml
@@ -16,6 +16,7 @@
     <explicitExcludes>
       <Path>.idea</Path>
       <Path>.vs</Path>
+      <Path>Library</Path>
       <Path>Temp</Path>
       <Path>obj</Path>
     </explicitExcludes>
