+++ Through the Dungeon/Assets/Animations/Characters/Player/Run_Down.anim
@@ -22,32 +22,32 @@ AnimationClip:
     - time: 0
       value: {fileID: -4590275256835052575, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.083333336
+    - time: 0.06666667
       value: {fileID: 7334058142527188712, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.16666667
+    - time: 0.13333334
       value: {fileID: -121124710891825258, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.25
+    - time: 0.2
       value: {fileID: -1163346415692482620, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.33333334
+    - time: 0.26666668
       value: {fileID: 3810578398183644174, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.41666666
+    - time: 0.33333334
       value: {fileID: 5416046831821694291, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.5
+    - time: 0.4
       value: {fileID: 8887354878593949365, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.5833333
+    - time: 0.46666667
       value: {fileID: -2599282569545456462, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
     script: {fileID: 0}
-  m_SampleRate: 12
+  m_SampleRate: 15
   m_WrapMode: 0
   m_Bounds:
     m_Center: {x: 0, y: 0, z: 0}
@@ -75,7 +75,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.6666666
+    m_StopTime: 0.53333336
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
