+++ Through the Dungeon/Assets/Animations/Characters/Player/Run_Left.anim
@@ -22,32 +22,32 @@ AnimationClip:
     - time: 0
       value: {fileID: -8046333585130505124, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.083333336
+    - time: 0.06666667
       value: {fileID: -2370035766628566219, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.16666667
+    - time: 0.13333334
       value: {fileID: 3113715368076444481, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.25
+    - time: 0.2
       value: {fileID: -2623049692062273706, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.33333334
+    - time: 0.26666668
       value: {fileID: 8516824271351505910, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.41666666
+    - time: 0.33333334
       value: {fileID: -8838169628664894647, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.5
+    - time: 0.4
       value: {fileID: -3847083532291991898, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
-    - time: 0.5833333
+    - time: 0.46666667
       value: {fileID: 1825421163732626159, guid: 3d8be72d5d3af674fa94127f97c7d94d,
         type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
     script: {fileID: 0}
-  m_SampleRate: 12
+  m_SampleRate: 15
   m_WrapMode: 0
   m_Bounds:
     m_Center: {x: 0, y: 0, z: 0}
@@ -75,7 +75,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.6666666
+    m_StopTime: 0.53333336
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
