+++ Through the Dungeon/Assets/Scripts/Plugins.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0c43aa2406d7460e9d507600417b5e98
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
